fix: include fees in journal entry amounts
